# api/openapi.yaml
openapi: 3.1.0
info:
  title: EVE-Cyno Data API
  summary: Deterministic EVE Online data, fit calculations and tools. No language model in the loop.
  description: |-
    Deterministic EVE Online data and calculations: community-fit search, fit breakdown and statistics (the Gofa dogma engine), SDE item search, and the tool layer behind the EVE-Cyno assistant, also reachable over MCP. Nothing here calls a language model; the same inputs give the same answers.

    ## Conventions

    - Errors are JSON, always: `{"error":{"code":"...","message":"..."}}`. The codes are stable (see `ErrorDetail`).
    - Every response carries `X-Request-ID`: your value when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one.
    - Requests are rate limited per caller (a sliding window; the budget is stated on each operation): per API key when one is sent, otherwise per client IP. A refused request is `429` with `Retry-After`.
    - Request bodies are capped (256 KiB on the fit routes, 64 KiB on tools and MCP); an oversize body is `413`.
...

    ## Attribution and licences

    © 2014 Fenris Creations hf. All rights reserved. 'EVE', 'EVE Online', 'Fenris Creations', and all related logos and images are trademarks or registered trademarks of Fenris Creations hf. This material is used with limited permission of Fenris Creations. No official affiliation or endorsement by Fenris Creations is stated or implied. Created under the EVE Developer License Agreement.

...
servers:
...
paths:
  /fits/search:
    get:
      tags:
        - fits
      summary: Search community fits
      description: |-
        Searches the community-fit corpus (EVE Workbench, Abysstracker and other sources named in each hit). The facet parameters filter the corpus; with `q` a filtered vector search ranks the hits. When the facets match nothing they are relaxed one at a time, last first: `relaxed` is then true and `dropped` names what was removed. `ship` may be a loose hull name (`megath` finds Megathron). No language model is involved. Every hit carries an `attribution` object (`source`, `source_url`, `author` when known, `license`, `credit` and `license_url` when the terms require them); show it wherever you show the fit. The fits are community data served ranked and capped, not for bulk export.

        Rate limit: 120 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: searchFits
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: ship
          in: query
          required: false
          description: Hull name; a partial or loose name is resolved against the SDE.
          schema:
            type: string
        - name: activity
          in: query
          required: false
          description: Activity tag, e.g. `pve` or `pvp`.
          schema:
            type: string
        - name: tag
          in: query
          required: false
          description: Extra tag, e.g. `abyss` or `ratting`.
          schema:
            type: string
        - name: filament
          in: query
          required: false
          description: Abyssal filament type.
          schema:
            type: string
        - name: cost
          in: query
          required: false
          description: Cost class tag, e.g. `cheap`.
          schema:
            type: string
        - name: source
          in: query
          required: false
          description: Restrict to one corpus source; empty searches the default source set.
          schema:
            type: string
        - name: clone
          in: query
          required: false
          description: '`alpha` keeps only alpha-clone fits, `omega` only omega-only fits; empty keeps both.'
          schema:
            type: string
            enum:
              - alpha
              - omega
        - name: q
          in: query
          required: false
          description: Free-text query; empty lists by the facets alone.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: 'Maximum hits, clamped to 1-24 (the response `limit` and `note` say so). There is no offset, page or cursor: one query reaches one ranked page, and a request carrying such a parameter is refused with `invalid_request`.'
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: The ranked fits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitSearchResult'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "429":
          $ref: '#/components/responses/RateLimited'
        "502":
          $ref: '#/components/responses/UpstreamError'
        "503":
          $ref: '#/components/responses/Unavailable'
  /fits/detail:
    post:
      tags:
        - fits
      summary: Describe a fit
      description: |-
        Parses an EFT block and returns its slots, CPU / powergrid / calibration use and ship class. Deterministic; DPS and EHP are not part of this view (see `/fit/stats`).

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: fitDetail
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitDetailRequest'
      responses:
        "200":
          description: The fit breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitDetail'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "502":
          $ref: '#/components/responses/UpstreamError'
        "503":
          $ref: '#/components/responses/Unavailable'
  /fit/stats:
    post:
      tags:
        - fits
      summary: Compute fit statistics
      description: |-
        Parses an EFT block and returns the Gofa engine's DPS, tank, capacitor, navigation, targeting and drone numbers. Deterministic; the same engine produces the numbers of the `compute_fit_stats` tool.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: fitStats
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitStatsRequest'
      responses:
        "200":
          description: The computed statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitStats'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "502":
          $ref: '#/components/responses/UpstreamError'
        "503":
          $ref: '#/components/responses/Unavailable'
  /fit/suggest:
    post:
      tags:
        - fits
      summary: Suggest modules for an empty slot
      description: |-
        Ranks the modules that can fill a slot tier of the given fit: how often similar community fits carry them (when the corpus is up), their meta tier, and whether they fit the remaining CPU and powergrid. Without the corpus the ranking falls back to the SDE alone.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: fitSuggest
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitSuggestRequest'
      responses:
        "200":
          description: At most 12 ranked suggestions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Suggestion'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "503":
          $ref: '#/components/responses/Unavailable'
  /items/search:
    get:
      tags:
        - items
      summary: Search modules, charges and drones
      description: |-
        Name search over the SDE's fittable modules, charges or drones (the fitting palette's search).

        Rate limit: 120 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: searchItems
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: q
          in: query
          required: true
          description: Name fragment; at least 2 characters.
          schema:
            type: string
            minLength: 2
        - name: kind
          in: query
          required: false
          description: What to search.
          schema:
            type: string
            enum:
              - module
              - charge
              - drone
            default: module
        - name: slot
          in: query
          required: false
          description: 'For modules: restrict to one slot tier.'
          schema:
            type: string
            enum:
              - high
              - mid
              - low
              - rig
        - name: limit
          in: query
          required: false
          description: Maximum hits.
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: The matching items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleHit'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "429":
          $ref: '#/components/responses/RateLimited'
        "503":
          $ref: '#/components/responses/Unavailable'
...
  /tool/compute_fit_stats:
    post:
      tags:
        - tools
      summary: 'Compute the real stats of an EVE Online fit from its EFT text with the dogma engine (all-V skills, no implants): DPS…'
      description: |-
        Compute the real stats of an EVE Online fit from its EFT text with the dogma engine (all-V skills, no implants): DPS per weapon group with the charge used, volley, EHP per layer with resists, max speed (propulsion on), align time, capacitor stability and CPU/PG load. Call this whenever the user asks for a fit's DPS, EHP, tank, speed, align or cap stability - never estimate these by hand. Weapons with no charge on their EFT line get a default T1 ammo (Scourge missiles; Phased Plasma / Antimatter / Multifrequency for turrets), which the output labels.

        Sources: EVE Static Data Export (via Fuzzwork). `result.data` is a `FitStatCard` object when the call succeeds.

        Access tier: **public**. No key needed.

        Rate limit: 30 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: compute_fit_stats
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                charges:
                  type: object
                  description: 'Optional ammo to use instead of the default, as {module name: charge name}, e.g. {"Rocket Launcher II": "Mjolnir Rocket"}. Only for weapons with no charge on their EFT line.'
                  additionalProperties:
                    type: string
                eft_text:
                  type: string
                  description: Full fitting in EFT format, starting with [Ship Name, Fit Name]. Loaded ammo may be given as 'Module, Charge' lines.
              required:
                - eft_text
      responses:
        "200":
          description: The tool's answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitStatCardToolResponse'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
        "503":
          $ref: '#/components/responses/Unavailable'
        "504":
          $ref: '#/components/responses/Timeout'
...
  /mcp:
    post:
      tags:
        - mcp
      summary: Model Context Protocol endpoint
      description: |-
        Streamable-HTTP MCP server (stateless, JSON responses): the same tools as `/tool/{name}`, for MCP clients such as Claude Desktop or Claude Code. POST one JSON-RPC 2.0 message per request (`initialize`, `tools/list`, `tools/call`, ...); a notification is answered `202` with no body. `GET` and `DELETE` are not supported in stateless mode (405). Point a client at this URL; it is not meant to be called by hand.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: mcp
      parameters:
        - $ref: '#/components/parameters/RequestID'
      security:
        - {}
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A JSON-RPC 2.0 request or notification (see https://modelcontextprotocol.io/specification).
              properties:
                jsonrpc:
                  const: "2.0"
                id:
                  type:
                    - string
                    - integer
                method:
                  type: string
                  examples:
                    - tools/list
                params:
                  type: object
              required:
                - jsonrpc
                - method
      responses:
        "200":
          description: The JSON-RPC 2.0 response.
          content:
            application/json:
              schema:
                type: object
                description: 'A JSON-RPC 2.0 response: `result` on success, `error` otherwise.'
                properties:
                  jsonrpc:
                    const: "2.0"
                  id:
                    type:
                      - string
                      - integer
                      - "null"
                  result:
                    type: object
                  error:
                    type: object
                required:
                  - jsonrpc
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
...
  /openapi.yaml:
    get:
      tags:
        - meta
      summary: This API description (YAML)
      description: |-
        The OpenAPI 3.1 document of this service, generated from its route table and tool registry.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: openapiYaml
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          description: The OpenAPI document.
          content:
            application/yaml:
              schema:
                type: string
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "429":
          $ref: '#/components/responses/RateLimited'
  /openapi.json:
    get:
      tags:
        - meta
      summary: This API description (JSON)
      description: |-
        The same OpenAPI 3.1 document as JSON.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: openapiJson
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          description: The OpenAPI document.
          content:
            application/json:
              schema:
                type: object
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "429":
          $ref: '#/components/responses/RateLimited'
components:
  schemas:
...
    FitSearchResult:
      type: object
      properties:
        total:
          type: integer
        relaxed:
          type: boolean
        dropped:
          type: array
          items:
            type: string
        hits:
          type: array
          items:
            $ref: '#/components/schemas/FitSearchHit'
        ship_counts:
          type: object
          additionalProperties:
            type: integer
        limit:
          type: integer
        note:
          type: string
      required:
        - total
        - relaxed
        - dropped
        - hits
        - limit
...
  responses:
    BadRequest:
      description: 'The request is malformed: invalid JSON, or a missing or invalid parameter or header (`invalid_request`, `janice_key_required`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
    RateLimited:
      description: The caller is over the rate limit; retry after the number of seconds in `Retry-After` (`rate_limited`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: An ID to correlate this request with the service's logs; echoed in the response.
      schema:
        type: string
        pattern: ^[A-Za-z0-9._-]{1,64}$
  headers:
    RequestID:
      description: 'The request''s ID: the caller''s `X-Request-ID` when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one. Quote it when reporting a problem.'
      schema:
        type: string
    RetryAfter:
      description: Seconds until the caller's budget has room again.
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      description: The same API key in the `X-API-Key` header, for clients that cannot set `Authorization`.
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      description: 'An API key as `Authorization: Bearer <key>`. Keyed tools need one; the other tools ignore it, but a key that is sent and not valid is refused (`401 invalid_api_key`).'
      scheme: bearer